requirements.txt
aiogram == 3.20.0
openpyxl == 3.1.5
pandas == 2.2.3
python-dotenv == 1.1.0
